+++ CHANGELOG.rst
@@ -3,6 +3,14 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_ after 2-1.
 
+`Unreleased`
+------------
+
+
+Added:
+
+- maxVelocityPercent added to motors. Allows turnaround times to be tweaked
+  without changing VMAX
 
 `4-0`_ - 2019-07-05
 -------------------